conditional rendering for I am a Vendor or not working
