+++ src/components/RegisterPage/RegisterPage.jsx
@@ -4,23 +4,23 @@ import { useHistory } from 'react-router-dom';
 import RegisterForm from '../RegisterForm/RegisterForm';
 
 // Material-UI
-import { ThemeProvider } from '@material-ui/core/styles';
 import {
   Button, // replaces html5 <button> element
   Grid,
   Typography, // replace html5 elements dealing with text, <h1>, <h2>, <h3>, <p>, etc...
 } from '@material-ui/core';
 
-function RegisterPage({ theme }) {
+function RegisterPage() {
   const history = useHistory();
 
   return (
-    <ThemeProvider theme={theme}>
+    <Grid>
 
       <Grid item container xs={12}>
         <RegisterForm />
       </Grid>
 
+      {/* Moved this Button to RegisterForm */}
       {/* <Grid item container xs={12} justify="center"> */}
       {/* <Button color="secondary" variant="contained"
           type="button"
@@ -31,7 +31,7 @@ function RegisterPage({ theme }) {
           Login
         </Button> */}
       {/* </Grid> */}
-    </ThemeProvider>
+    </Grid>
   );
 }
